Add virtual key option in KeyCommand
Introduce a new "-v" flag to support virtual key distribution in the KeyCommand. This allows keys to be given virtually, bypassing physical item distribution, which is especially useful for offline players. The changes ensure that keys can be distributed both as items and virtually, based on the presence of the "-v" flag.